tests reorganize
