+++ test_report/tests/common/mod.rs
@@ -1,0 +1,5 @@
+//! Shared helpers for `test_report` integration tests.
+
+mod env;
+
+pub use env::{clean_up_cache_files, cleanup_env_vars};
